getCurrentTimeMS now returns long

+++ app/src/main/java/com/lucek/androidgameengine2d/core/graphics/Window.java
@@ -39,6 +39,6 @@ public class Window {
         return this.camera;
     }
 
-    public float getCurrentTimeMS() { return System.currentTimeMillis(); }
+    public long getCurrentTimeMS() { return System.currentTimeMillis(); }
 
 }
